requirements.txt
playsound==1.3.0
PyAudio==0.2.11
RPi.GPIO==0.7.1
SpeechRecognition==3.8.1
